requirements.txt
# Core
python-dotenv==1.0.1

# Data Engineering
duckdb==0.10.3
dbt-duckdb==1.8.1
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0
sqlalchemy==2.0.30

# Data Quality
great-expectations==0.18.19

# ML / AI
scikit-learn==1.5.0
xgboost==2.0.3
torch==2.3.1
shap==0.45.1
mlflow==2.13.2
boto3==1.34.131

# GenAI / LLM
langchain==0.2.5
langchain-openai==0.1.9
langchain-aws==0.1.6
langsmith==0.1.77
openai==1.35.3

# Orchestration
apache-airflow==2.9.2

# API / Serving
fastapi==0.111.0
uvicorn==0.30.1

# Visualisation / UI
streamlit==1.36.0
plotly==5.22.0
matplotlib==3.9.0

# Testing / Linting
pytest==8.2.2
pytest-cov==5.0.0
ruff==0.4.9
faker==25.8.0
